requirements.txt
django==5.2.6
djangorestframework
djangorestframework-simplejwt
dj-rest-auth
django-allauth
social-auth-app-django
django-cors-headers
